// src/types/models.d.ts
export interface QuoteShift {
  id: string;
  quote_id: string;
  quoteId?: string;
  day: Day;
  start_time: string;
  startTime?: string;
  end_time: string;
  endTime?: string;
  break_duration: number;
  breakDuration?: number;
  number_of_cleaners: number;
  numberOfCleaners?: number;
  employment_type: EmploymentType;
  employmentType?: EmploymentType;
  level: EmployeeLevel;
  allowances: string[];
  estimated_cost: number;
  estimatedCost?: number;
  location: string;
  notes: string;
}